requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.1
numpy>=1.26.0
pandas>=2.2.0
openpyxl>=3.1.3

# LangChain / LangGraph
langchain>=0.2.0
langchain-core>=0.2.0
langchain-community>=0.2.0
langchain-openai>=1.0.0
langchain-huggingface>=0.1.0
langchain-chroma>=0.1.0
langgraph>=1.0.3
langgraph-prebuilt>=1.0.4

# Retrieval / embedding
openai>=1.50.0
huggingface-hub>=0.24.0
chromadb>=0.5.3
sentence-transformers>=2.7.0
rank-bm25>=0.2.2

# Parsing
pdfplumber>=0.11.0
pymupdf>=1.23.8
python-docx>=1.1.2
olefile>=0.47
camelot-py[cv]>=0.11.0
pdf2image>=1.17.0
pillow>=10.3.0

# Optional OCR/QA (이미지 포함 문서일 때)
paddleocr>=2.7.0

# FastAPI 업로드용
python-multipart>=0.0.9

# Utilities / eval / logging
requests>=2.31.0
tqdm>=4.66.0
wandb>=0.16.0
langsmith>=0.1.29
scikit-learn>=1.4.0

# 선택: 오프라인 BM25/FAISS 실험
faiss-cpu>=1.7.4
